Fixing the error

+++ Program.cs
@@ -2,8 +2,7 @@
 var part = "Part 7";
 Console.WriteLine($"Hello, World! {part}");
 
-
 int a = int.MaxValue;
 int b = 1;
-long c = a + b;
+long c = (long)a + b;
 Console.WriteLine(c);
